+++ sentry/src/main/java/io/sentry/NoOpHub.java
@@ -180,6 +180,11 @@ public final class NoOpHub implements IHub {
   }
 
   @Override
+  public @Nullable ITransaction getTransaction() {
+    return null;
+  }
+
+  @Override
   public @NotNull SentryOptions getOptions() {
     return emptyOptions;
   }