+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -918,10 +918,16 @@ public final class Sentry {
   /**
    * Gets the current active transaction or span.
    *
-   * @return the active span or null when no active transaction is running
+   * @return the active span or null when no active transaction is running. In case of
+   *     globalHubMode=true, always the active transaction is returned, rather than the last active
+   *     span.
    */
   public static @Nullable ISpan getSpan() {
-    return getCurrentHub().getSpan();
+    if (globalHubMode) {
+      return getCurrentHub().getTransaction();
+    } else {
+      return getCurrentHub().getSpan();
+    }
   }
 
   /**